# runtime/packages/client/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Scrapyomama Runtime API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
# Par défaut : session d'interface (cookie) ou clé d'API (Bearer, scopes de 13 § 8). Les routes publiques posent `security: []` ;
# les routes réservées à la session (registre `auth: 'session'`) posent `security: [{sessionCookie: []}]` : jamais une clé d'API
# sur une route d'administration (13 § 8). Le test de dérive compare ce mode d'authentification au registre du serveur.
security:
  - sessionCookie: []
  - apiKey: []
...
paths:
...
  /api/events:
    get:
      operationId: streamEvents
      tags: [events]
      summary: Flux SSE multiplexé, un seul par onglet (06 § 3)
      description: >-
        Trames `id:`, `event:`, `data:` (JSON), commentaire `: ping` toutes les 15 à 20 s. Reprise par l'en-tête
        `Last-Event-ID` sur des événements persistés : la console le renvoie à chaque reconnexion et ignore tout
        événement dont l'identifiant a déjà été vu. Les noms d'événements sont définis avec le serveur (tâche 3.1).
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Flux d'événements.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
...
  /api/runs/{id}/events:
    get:
      operationId: streamRunEvents
      tags: [events]
      summary: Vue filtrée du flux SSE pour un run ou une enquête, rejouée depuis `investigation_events`
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        '429':
          $ref: '#/components/responses/Error'
        '200':
          description: Flux d'événements (mêmes trames que `/api/events`).
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
  /api/datasets/{id}/items:
    get:
      operationId: getDatasetItems
      tags: [datasets]
      summary: Items d'un dataset, exportés en flux (JSON, NDJSON, CSV à cellules neutralisées), curseur `after`
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: format
          in: query
          schema:
            type: string
            enum: [json, ndjson, csv]
            default: json
        - name: after
          in: query
          description: Curseur opaque du dernier item servi (`next_cursor`, ou en-tête `X-Next-Cursor`).
          schema:
            type: string
        - name: limit
          in: query
          description: Nombre maximal d'items ; absent, l'export va jusqu'au bout du dataset, en flux.
          schema:
            type: integer
            minimum: 1
            maximum: 100000
        - name: fields
          in: query
          description: Champs à garder, séparés par des virgules.
          schema:
            type: string
        - name: omit
          in: query
          description: Champs à retirer, séparés par des virgules.
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: >-
            Items, écrits en flux (mémoire bornée) ; pièce jointe avec `nosniff` pour un export. CSV : cellules qui
            commencent par =, +, -, @, tabulation ou retour chariot préfixées d'une apostrophe (08b § 2).
          headers:
            X-Next-Cursor:
              description: Curseur de la suite quand `limit` arrête la page avant la fin du dataset.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetItems'
            application/x-ndjson:
              schema:
                type: string
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sy.session
      description: Session d'interface (13 § 5) ; `sy.session` sans le préfixe `__Host-` quand PUBLIC_URL est en http (développement).
    apiKey:
      type: http
      scheme: bearer
      description: Clé d'API `sy_live_…` à scopes (13 § 8) ; jamais de scope d'administration.
    deviceToken:
      type: http
      scheme: bearer
      description: Jeton d'appareil de l'extension (07 § 1), lié à (utilisateur, appareil) ; jamais accepté hors de /api/extension.
    metricsToken:
      type: http
      scheme: bearer
      description: METRICS_TOKEN (14 § 3), distinct des clés d'API.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: "Erreur au format commun `{ error: { code, message } }` (le code est stable, jamais localisé)."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    DatasetItems:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        next_cursor:
          type: [string, 'null']
    # ——— Stratégies et statut ———
...